feat: enhance note update and removal with authorization checks

+++ backend/src/notes/notes.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common';
 import { CreateNoteDTO } from './dto/create-note.dto';
 import { UpdateNoteDTO } from './dto/update-note.dto';
 import { Note } from './entities/note.entity';
@@ -42,16 +42,20 @@ export class NotesService {
     return note;
   }
 
-  async update(id: string, dto: UpdateNoteDTO) {
+  async update(id: string, dto: UpdateNoteDTO, payload: JwtPayload) {
     const updatedNote = await this.notesRepository.preload({ id, ...dto });
 
     if (!updatedNote) this.throwNotFoundException();
 
+    if (payload.sub !== updatedNote.user.id) throw new UnauthorizedException(`You can't change another user note.`);
+
     return await this.notesRepository.save(updatedNote);
   }
 
-  async remove(id: string) {
+  async remove(id: string, payload: JwtPayload) {
     const note = await this.findOne(id);
+
+    if (payload.sub !== note.user.id) throw new UnauthorizedException(`You can't change delete user note.`);
 
     return await this.notesRepository.remove(note);
   }
